Bug 1475689 - don't wait for commands to complete with https://golang.org/pkg/os/exec/#Cmd.Wait
Calling https://golang.org/pkg/os/exec/#Cmd.Wait can mean that you wait
forever, if a process gets stuck and can't be killed when the task is
aborted. Instead signal task abortion via closing of a go channel so
that task abortion can be handled in parallel to process termination.
This was already implemented on Windows, this applies a similar fix for
all other platforms.