Bug 1436195 - Allow exposing a specific port with stateless DNS
This moves the livelog process to a specific "internal" port, and then
allows using a predefined port for exposures, which is necessary when
there is only one port open in the firewall.

This means that stateless DNS can only support exposing one thing at a
time -- but that was already the case.

Once stateless DNS is no longer supported, it doesn't make sense for the
livelog get *or* put ports be part of the configuration.  In fact, at
some point they will need to be dynamically generated to support running
multiple livelog processes at the same time.  In the interim, while
stateless DNS is still supported, the configuration looks a little funny
but (critically) retains compatibility.