Bug 1536780 - Add contentEncoding to task payload (#239)
It is now possible to explicitly declare the Content-Encoding of file or
directory artifacts inside the task payload, via the `contentEncoding`
property. If not provided, the content encoding will be chosen by the
worker based on the file extension of the artifact, as before. If
specified for a directory artifact, the content encoding will apply to
all files inside the directory.